# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aiome Management Console API",
    "description": "Core API for the Autonomous AI Operating System",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/commune/autonomous/start": {
      "post": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "autonomous_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAutonomousRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Started autonomous mode",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/commune/autonomous/status": {
      "get": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "autonomous_status",
        "responses": {
          "200": {
            "description": "Autonomous status",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/commune/autonomous/stop": {
      "post": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "autonomous_stop",
        "responses": {
          "200": {
            "description": "Stopped autonomous mode",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/commune/list": {
      "get": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "list_messages",
        "responses": {
          "200": {
            "description": "List recent messages",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/commune/send": {
      "post": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "send_message",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCommuneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/commune/status": {
      "get": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "commune_status",
        "responses": {
          "200": {
            "description": "Status retrieved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/commune/topics": {
      "get": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "list_topics",
        "responses": {
          "200": {
            "description": "Topics retrieved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "crate::routes::commune"
        ],
        "operationId": "create_topic",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Topic created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SendCommuneRequest": {
        "type": "object",
        "required": [
          "recipient_pubkey",
          "topic_id",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "recipient_pubkey": {
            "type": "string"
          },
          "topic_id": {
            "type": "string"
          }
        }
      },
...
      "StartAutonomousRequest": {
        "type": "object",
        "required": [
          "topic_id",
          "peer_pubkey"
        ],
        "properties": {
          "interval_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "max_rounds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "peer_pubkey": {
            "type": "string"
          },
          "topic_id": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_SERVER_SECRET"
      }
    }
  }
}